+++ src/header/Stash StashID Icon.user.js
@@ -1,0 +1,12 @@
+// ==UserScript==
+// @name        Stash StashID Icon
+// @namespace   %NAMESPACE%
+// @description Adds checkmark icon to performer and studio cards that have a stashid
+// @version     0.1.0
+// @author      7dJx1qP
+// @match       %MATCHURL%
+// @grant       unsafeWindow
+// @grant       GM_addStyle
+// @require     %LIBRARYPATH%
+// @require     %FILEPATH%
+// ==/UserScript==